rm weird lifetime / reference versions of std::ops
