+++ jolt-core/src/jolt/instruction/slt.rs
@@ -37,7 +37,7 @@ impl JoltInstruction for SLTInstruction {
 
         for (ltu_i, eq_i) in ltu.iter().zip(eq) {
             ltu_sum += *ltu_i * eq_prod;
-            eq_prod *= eq_i;
+            eq_prod *= *eq_i;
         }
 
         // x_s * (1 - y_s) + EQ(x_s, y_s) * LTU(x_{<s}, y_{<s})
