+++ jolt-core/src/r1cs/spartan.rs
@@ -272,7 +272,7 @@ impl<F: JoltField, C: CommitmentScheme<Field = F>> UniformSpartanProof<F, C> {
                 .zip(poly_Cz.evals_ref())
                 .enumerate()
             {
-                if *az * bz != *cz {
+                if *az * *bz != *cz {
                     let padded_segment_len = segmented_padded_witness.segment_len;
                     let error_segment_index = i / padded_segment_len;
                     let error_step_index = i % padded_segment_len;
@@ -350,7 +350,7 @@ impl<F: JoltField, C: CommitmentScheme<Field = F>> UniformSpartanProof<F, C> {
             let compute_eval_table_sparse_single = |small_M: &Vec<(usize, usize, F)>| -> Vec<F> {
                 let mut small_M_evals = vec![F::zero(); key.shape_single_step.num_vars + 1];
                 for (row, col, val) in small_M.iter() {
-                    small_M_evals[*col] += eq_rx_con[*row] * val;
+                    small_M_evals[*col] += eq_rx_con[*row] * *val;
                 }
                 small_M_evals
             };
