+++ rust-toolchain.toml
@@ -1,4 +1,4 @@
 [toolchain]
-channel = "nightly-2023-09-22"
+channel = "nightly-2024-04-29"
 targets = ["riscv32i-unknown-none-elf"]
